% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsConvertACF.R
\name{tsConvertACF}
\alias{tsConvertACF}
\title{tsConvertACF}
\usage{
tsConvertACF(data = rnorm(30, 10, 2))
}
\arguments{
\item{data}{Data for analysis}
}
\value{
ACF of time series of the data
}
\description{
Returns ACF of time series of the data
}
\examples{
tsConvertACF()
}
