% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsConvertCUSUM.R
\name{tsConvertCUSUM}
\alias{tsConvertCUSUM}
\title{tsConvertCUSUM}
\usage{
tsConvertCUSUM(data = rnorm(30, 10, 2))
}
\arguments{
\item{data}{Data for analysis}
}
\value{
CUSUM plot of an auto.arima fitted model of the data
}
\description{
tsConvertCUSUM
}
\examples{
tsConvertCUSUM()
}
