% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsConvertPACF.R
\name{tsConvertPACF}
\alias{tsConvertPACF}
\title{tsConvertPACF}
\usage{
tsConvertPACF(data = rnorm(30, 10, 2))
}
\arguments{
\item{data}{Data for analysis}
}
\value{
PACF of time series of the data
}
\description{
tsConvertPACF
}
\examples{
tsConvertPACF
}
