% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsConvertPlot.R
\name{tsConvertPlot}
\alias{tsConvertPlot}
\title{tsConvertPlot}
\usage{
tsConvertPlot(data = rnorm(30, 10, 2))
}
\arguments{
\item{data}{Data for analysis}
}
\value{
A plot of the data in time series format
}
\description{
Creates a time series plot of the data
}
\examples{
tsConvertPlot()
}
