% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsConvertAccuracy.R
\name{tsConvertAccuracy}
\alias{tsConvertAccuracy}
\title{tsConvertAccuracy}
\usage{
tsConvertAccuracy(data = rnorm(30, 10, 2))
}
\arguments{
\item{data}{Data for analysis}
}
\value{
Accuracy of an auto.arima fitted model
}
\description{
tsConvertAccuracy
}
\examples{
tsConvertAccuracy()
}
